move around and remove uneeded files
