--- addon/utils/computed/is-empty-object.ts
@@ -1,16 +1,0 @@
-import { assert } from '@ember/debug';
-import { computed, get } from '@ember/object';
-import ComputedProperty from '@ember/object/computed';
-import { isPresent } from '@ember/utils';
-
-const { keys } = Object;
-
-export default function isEmptyObject(
-  dependentKey: string
-): ComputedProperty<unknown, unknown> {
-  assert('`dependentKey` must be defined', isPresent(dependentKey));
-
-  return computed(dependentKey, function() {
-    return keys(get(this, dependentKey)).length === 0;
-  }).readOnly();
-}
